requirements.txt
flask==3.0.3
openpyxl==3.1.2
requests==2.31.0
pandas==2.0.2
numpy==1.26.4
python-dotenv==1.0.1
